// src/app/Models/episode-model.ts
export class EpisodeModel {
    EpisodeId: number;
    EpisodeTitle: string;
    EpisodeDescription: string;
    EpisodeUrl: string;
    EpisodeImg: string;
    EpisodeVisible: boolean;
    CreationDate: Date;
    EpisodeViews: number;
    EpisodePublishDate: string;
    SessionId: number;
    SeasonTitle:string;
    ProgramId: number;
    ProgramName: string;
    CategoryId: number;
    CategoryTitle: string;
    ProgramTypeId: number;
    ProgramTypeTitle: string;
    InterviewerName:string;
    Hour:any;
    Minute:any;
    Date:string;
}